# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Karo Online API
  version: 1.0.0
  description: Production Express API ported from TanStack server functions.
servers:
  - url: http://localhost:4000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /v1/admin/users/{userId}:
    get:
      summary: Full user dossier
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: OK }
  /v1/admin/users/{userId}/customer:
    patch:
      summary: Patch customer profile
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: OK }
  /v1/admin/users/{userId}/vendor:
    patch:
      summary: Patch vendor profile
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: OK }
  /v1/admin/users/{userId}/wallet:
    post:
      summary: Credit or debit wallet
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: OK }
  /v1/admin/users/{userId}/block:
    post:
      summary: Block or unblock user
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: OK }
  /v1/admin/vendors/{userId}/approval:
    post:
      summary: Approve or unapprove vendor
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: OK }
  /v1/admin/kyc/{id}/status:
    post:
      summary: Set KYC status
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: OK }
  /v1/admin/table/{table}:
    get:
      summary: Generic admin list
      parameters:
        - in: path
          name: table
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
    post:
      summary: Generic admin create
      parameters:
        - in: path
          name: table
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
    patch:
      summary: Generic admin update
      parameters:
        - in: path
          name: table
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
    delete:
      summary: Generic admin delete
      parameters:
        - in: path
          name: table
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
...